requirements.txt
Django==1.8.7
django-braces==1.8.1
django-crispy-forms==1.5.2
django-registration-redux==1.2
Pillow==3.1.0
psycopg2==2.6.1
six==1.10.0
wheel==0.24.0
